use AllowEmptyClientSecret config setting
to reduce confusion we do not allow an empty client secret by default,
however a user can set `AllowEmptyClientSecret` to true which will allow
it, but will print out a warning message in the logs.

Clarity provided in the example configmap which appears to be the
current defacto documentation for config settings.

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>